refactor: reorganize subtitle-fast-comparator module structure
- Restructure codebase for better organization:
  - Move comparison algorithms to comparators/ directory
    - chamfer_edge.rs
    - hybrid_mask.rs
    - spectral_hash.rs
    - structural_dssim.rs
  - Move pipeline utilities to pipeline/ directory
    - feature.rs (feature extraction)
    - ops.rs (operations)
    - preprocess.rs (preprocessing)

- Update factory.rs to use new module paths
- Update lib.rs with reorganized module structure
- Update tests.rs to reflect new organization
- Add mod.rs files for new directories

This refactoring improves code maintainability and makes the
module structure more intuitive by grouping related functionality.
