+++ crates/subtitle-fast-comparator/src/comparators/structural_dssim.rs
@@ -1,10 +1,12 @@
 use subtitle_fast_decoder::YPlaneFrame;
 use subtitle_fast_validator::subtitle_detection::RoiConfig;
 
-use crate::SubtitleComparator;
-use crate::feature::{ComparisonReport, FeatureBlob, ReportMetric};
-use crate::ops::resize_average;
-use crate::preprocess::{MaskedPatch, PreprocessSettings, extract_masked_patch};
+use crate::comparators::SubtitleComparator;
+use crate::pipeline::ops::resize_average;
+use crate::pipeline::preprocess::extract_masked_patch;
+use crate::pipeline::{
+    ComparisonReport, FeatureBlob, MaskedPatch, PreprocessSettings, ReportMetric,
+};
 
 const TAG: &str = "structural-dssim";
 const MAX_DIM: usize = 64;
