+++ crates/subtitle-fast-comparator/src/pipeline/mod.rs
@@ -1,0 +1,6 @@
+pub mod feature;
+pub mod ops;
+pub mod preprocess;
+
+pub use feature::{ComparisonReport, FeatureBlob, ReportMetric};
+pub use preprocess::{MaskedPatch, PreprocessSettings};
